fix(screener): id BigInteger + last_updated nullable=False per spec
- Changed TickerSnapshot.id from Integer to BigInteger in models.py
- Changed id column from sa.Integer() to sa.BigInteger() in migration
- Added nullable=False to last_updated column in models.py (matches migration)
- Updated tests to manually set IDs (SQLite + BigInteger autoincrement workaround)

All 4 tests pass. Migration down/up verified clean.

+++ app/backend/alembic/versions/d4e8a2c1b9f6_add_ticker_snapshots.py
@@ -25,7 +25,7 @@ depends_on: Union[str, Sequence[str], None] = None
 def upgrade() -> None:
     op.create_table(
         "ticker_snapshots",
-        sa.Column("id", sa.Integer(), primary_key=True, autoincrement=True),
+        sa.Column("id", sa.BigInteger(), primary_key=True, autoincrement=True),
         sa.Column("ticker", sa.String(length=20), nullable=False),
         sa.Column("market", sa.String(length=8), nullable=False),
         sa.Column("snapshot_date", sa.Date(), nullable=False),
